% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biomart}
\alias{biomart}
\title{Main function to query the biomart database}
\usage{
biomart(genes, mart, dataset, attributes, filter, ...)
}
\arguments{
\item{genes}{a character vector storing the gene ids of a organisms of interest to be queried against biomart.}

\item{mart}{a character string specifying the mart to be used, e.g. mart = "ensembl".}

\item{dataset}{a character string specifying the dataset within the specified mart that shall be used, e.g. dataset = "hsapiens_gene_ensembl".}

\item{attributes}{}

\item{filter}{}

\item{...}{}
}
\description{
This function takes a set of gene ids and the biomart specifications
and performs a biomart query for the given set of gene ids.
}
\author{
Hajk-Georg Drost
}

