% Generated by roxygen2 (4.0.2): do not edit by hand
\name{biomart_organisms}
\alias{biomart_organisms}
\title{Function to check for available biomart connections for a query organism}
\usage{
biomart_organisms(organism = NULL, update = FALSE)
}
\arguments{
\item{organism}{a character string specifying the scientific name of a query organism.
Default is \code{organism} = \code{NULL}. In this case all available biomart connections are returned.}
}
\description{
This function returns either all available biomart connections for
all available organisms for which biomart access is possible, or (when specified) returns all
organism specific biomart connections.
}
\details{
This function collects all available biomart connections and returns a table storing
the organism for which biomart connections are available as well as the corresponding mart and dataset.
}
\note{
When running the function for the first time, data retrieval will take a while,
due to the remote access to biomart. The result is then saved in a *.txt file named
"_biomart/listDatasets.txt" allowing subsequent queries to perform much faster.
}
\examples{
\dontrun{

# returning all available biomart connections
head(biomart_organisms(), 20)

# retrieving all available datasets and biomart connections for
# a specific query organism (scientific name)
biomart_organisms(organism = "Arabidopsis thaliana")

# you can also update the downloaded version using the "update = TRUE" argument
head(biomart_organisms(update = TRUE), 20)

}
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://biomart.org/}

Mapping identifiers for the integration of genomic datasets with the
R/Bioconductor package biomaRt. Steffen Durinck, Paul T. Spellman, Ewan
Birney and Wolfgang Huber, Nature Protocols 4, 1184-1191 (2009).

BioMart and Bioconductor: a powerful link between biological databases and
microarray data analysis. Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean
Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21,
3439-3440 (2005).
}
\seealso{
\code{\link[biomaRt]{listMarts}}, \code{\link[biomaRt]{listDatasets}}, \code{\link{biomart}}
}

