% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cachedir}
\alias{cachedir}
\title{Get directory to store back end files like kingdom summaries etc}
\usage{
cachedir(non_temp_cache = "~/.biomartr_cache_dir.rds")
}
\arguments{
\item{non_temp_cache}{"~/.biomartr_cache_dir.rds",}
}
\value{
reads the rds file, and returns the path for local cache, if not
existing, use tempdir().
}
\description{
Get directory to store back end files like kingdom summaries etc
}
\examples{
cachedir()
}
\seealso{
Other cachedir: 
\code{\link{cachedir_set}()}
}
\concept{cachedir}
