% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cachedir_set}
\alias{cachedir_set}
\title{Set directory to store back end files like kingdom summaries etc}
\usage{
cachedir_set(path)
}
\arguments{
\item{path}{the path to cache dir, example "~/Bio_data/biomartr_cache/"}
}
\value{
invisible(NULL), only save the file to path location
}
\description{
Set directory to store back end files like kingdom summaries etc
}
\examples{
# By default it is tmpdir()
cachedir()
# cachedir_set("~/Bio_data/biomartr_cache/")
cachedir()
}
\seealso{
Other cachedir: 
\code{\link{cachedir}()}
}
\concept{cachedir}
