% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ensembl.info.R
\name{ensembl_divisions}
\alias{ensembl_divisions}
\title{List all available ENSEMBL divisions}
\usage{
ensembl_divisions()
}
\description{
Retrieve a list of available databases on ENSEMBL for which \code{get.ensembl.info} can be retrieved.
}
\examples{
ensembl_divisions()
}
\author{
Hajk-Georg Drost
}
