% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evalAlignment}
\alias{evalAlignment}
\title{Function to quantify the statistical significance of a given pairwise alignment}
\usage{
evalAlignment(seq, subject, sample_size, FUN, ..., fit_distr = "norm",
  comp_cores = 1)
}
\arguments{
\item{seq}{a character vector storing a sequence as string for which random sequences shall be computed.}

\item{subject}{a character vector storing a subject sequence as string to which \code{seq} shall be pairwise aligned.}

\item{sample_size}{a numeric value specifying the number of random sequences that shall be returned.}

\item{FUN}{a pairwise alignment function such as \code{\link[Biostrings]{pairwiseAlignment}} or any other function
that takes sequence arguments as first and second input.}

\item{...}{additional arguments that shall be passed to \code{FUN}.}

\item{fit_distr}{a character string specifying the probability distribution that shall be fitted to the histogram
of scores returned by \code{\link{randSeqDistr}}. See \code{\link[fitdistrplus]{fitdist}} when method = "mme" for details.}

\item{comp_cores}{a numeric value specifying the number of cores you want to use for multicore processing.}
}
\value{
a p-value quantifying the statistical significance of the pairwise alignment of the input sequences.
}
\description{
This function allows to quantify the statistical significance of a given pairwise alignment
between a query and subject sequence based on a sampled score distribution returned by \code{\link{randSeqDistr}}.
}
\details{
The test statistic is developed using moment matching estimation of a given probability distribution that
is fitted to the alignment score vector returned by \code{\link{randSeqDistr}}. The corresponding distribution
parameters are estimated by the \code{\link[fitdistrplus]{fitdist}} and the p-value quantifying the statistical
significance of the pairwise alignment of the input sequences is returned.
}
\examples{
seq_example <- "MEDQVGFGF"
subject_example <- "AYAIDPTPAF"

p_val_align <- evalAlignment(seq_example, subject_example, 10,
                             Biostrings::pairwiseAlignment,
                             scoreOnly=TRUE, fit_distr = "norm",
                             comp_cores = 1)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{randSeqDistr}}, \code{\link{randomSeqs}}, \code{\link[fitdistrplus]{fitdist}}
}

