% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flymine.organisms.R
\name{flymine.organisms}
\alias{flymine.organisms}
\title{Retrieve All Organism Names Stored on FlyMine}
\usage{
flymine.organisms(organism = "all")
}
\arguments{
\item{organism}{a character string specifying the scientific name of an organism for which data information shall be retrieved.
Default is \code{organism} = \code{"all"}.}
}
\description{
This function extracts all organism names (scientific names) for which genomes,
proteomes, and CDS files are stored on the FlyMine server.
}
\details{
This function queries the FlyMine API and retrieves all organisms specific information
stored on FlyMine This way annotation information can be accessed allowing a better
transparency of genome data used for omics studies.
}
\examples{
# retrieve data for all available organisms stored in FlyMine
head(flymine.organisms())

# retrieve data for Arabidopsis thaliana available organisms stored in FlyMine
flymine.organisms(organism = "Drosophila melanogaster")
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://www.flymine.org/}
}

