% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geneSequence}
\alias{geneSequence}
\title{Function to retrieve biological sequences of a given set of genes}
\usage{
geneSequence(genes, db)
}
\arguments{
\item{genes}{a character vector storing the gene id of a organisms of interest to be queried against the
acnucdb database.}

\item{db}{}
}
\description{
This function takes an character vector storing gene ids of interest
and returns the biological sequence of the corresponding gene ids.
}
\details{
Sequence information is retrieved from the acnucdb database.
}
\author{
Hajk-Georg Drost
}

