% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getENSEMBLGENOMESInfo.R
\name{getENSEMBLGENOMESInfo}
\alias{getENSEMBLGENOMESInfo}
\title{Retrieve ENSEMBLGENOMES info file}
\usage{
getENSEMBLGENOMESInfo()
}
\description{
Retrieve species and genome information from http://rest.ensemblgenomes.org/info/species?content-type=application/json/.
}
\author{
Hajk-Georg Drost
}

