% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGenome}
\alias{getGenome}
\title{Function for downloading a specific genome stored on the NCBI ftp:// server.}
\usage{
getGenome(db = "refseq", kingdom = "plant", organism)
}
\arguments{
\item{db}{a character string specifying the database from which the genome shall be retrieved: 'refseq','genebank', or 'all'.}

\item{kingdom}{a character string specifying the kingdom of the organisms of interest,
e.g. "archaea","bacteria", "fungi", "invertebrate", "plant", "protozoa", "vertebrate_mammalian", or "vertebrate_other"}

\item{organism}{a character string specifying the scientific name of the organism of interest, e.g. 'Arabidopsis_thaliana'.}
}
\value{
A data.table storing the geneids in the first column and the DNA dequence in the second column.
}
\description{
This function retrieves a fasta-file storing the genome of an organism of interest and stores
the genome file in the folder '_ncbi_downloads/genomes'.
}
\details{
Internally this function loads the the overview.txt file from NCBI:

 refseq: ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/

 genebank: ftp://ftp.ncbi.nlm.nih.gov/genomes/genebank/

 all: ftp://ftp.ncbi.nlm.nih.gov/genomes/all/

and creates a directory '_ncbi_downloads/genomes' to store
the genome of interest as fasta file for future processing.
In case the corresponding fasta file already exists within the
'_ncbi_downloads/genomes' folder and is accessible within the workspace,
no download process will be performed.
}
\examples{
\dontrun{
# download the genome of Arabidopsis thaliana

}
}
\author{
Hajk-Georg Drost
}
\references{
ftp://ftp.ncbi.nlm.nih.gov/genomes/
}

