% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGENOMEREPORT.R
\name{getGENOMEREPORT}
\alias{getGENOMEREPORT}
\title{Retrieve NCBI GENOME_REPORTS file}
\usage{
getGENOMEREPORT(
  local_file = file.path(cachedir(), "_ncbi_downloads", "overview.txt")
)
}
\arguments{
\item{local_file}{character, file path, default: file.path(cachedir(), "_ncbi_downloads", "overview.txt")}
}
\value{
a tibble object with the report
}
\description{
Retrieves NCBI GENOME_REPORTS file from
ftp://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/overview.txt.
}
\examples{
\dontrun{
report <- getGENOMEREPORT()
report
}
}
\author{
Hajk-Georg Drost
}
