% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKingdoms.R
\name{getKingdoms}
\alias{getKingdoms}
\title{Retrieve available kingdoms of life stored in RefSeq}
\usage{
getKingdoms()
}
\description{
A short list of kingdoms of life that are stored in the RefSeq
database and that can be downloaded using e.g. \code{\link{meta.retrieval}}, \code{\link{getGenome}}, etc.
}
\examples{
getKingdoms()
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{meta.retrieval}}, \code{\link{getGenome}}
}

