% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMarts.R
\name{getMarts}
\alias{getMarts}
\title{Retrieve All Available BioMart Databases}
\usage{
getMarts()
}
\description{
This funcion queries the BioMart Interface and returns a table
storing all available BioMart databases.
}
\examples{
# get a table of all available databases from BioMart
getMarts()
}
\author{
Hajk-Georg Drost
}

