% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaGenomeSummary.R
\name{getMetaGenomeSummary}
\alias{getMetaGenomeSummary}
\title{Helper function to retrieve the assembly_summary.txt file from NCBI genbank metagenomes}
\usage{
getMetaGenomeSummary()
}
\description{
Retrieval function of the assembly_summary.txt file from NCBI genbank metagenomes.
}
\examples{
\dontrun{
test <- getMetaGenomeSummary()
test
} 
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{getKingdomAssemblySummary}}, \code{\link{getSummaryFile}}
}

