% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniProtInfo.R
\name{getUniProtInfo}
\alias{getUniProtInfo}
\title{Get uniprot info from organism}
\usage{
getUniProtInfo(organism, path = cachedir(), update = TRUE)
}
\arguments{
\item{organism}{character, name of organism}
}
\description{
Get uniprot info from organism
}
