% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniProtSTATS.R
\name{getUniProtSTATS}
\alias{getUniProtSTATS}
\title{Retrieve UniProt Database Information File (STATS)}
\usage{
getUniProtSTATS(update = FALSE)
}
\arguments{
\item{update}{shall the internal \code{\link{cachedir}} file be deleted and the \code{STATS}
file freshly downloaded from the UniProt FTP servers?}
}
\description{
The UniProt stores a \code{STATS} file to summarise all available information for their reference proteomes.
Users can now download this file and process it with \code{biomartr}.
}
\examples{
\dontrun{
# retrieve STATS file from UniProt
uniprot_info <- getUniProtSTATS(update = TRUE)
# look at results
uniprot_info
}
}
\author{
Hajk-Georg Drost
}
