% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMetaGenomes.R
\name{listMetaGenomes}
\alias{listMetaGenomes}
\title{List available metagenomes on NCBI Genbank}
\usage{
listMetaGenomes()
}
\description{
List available metagenomes on NCBI genbank. NCBI genbank allows users
to download entire metagenomes of several metagenome projects. This function lists
all available metagenomes that can then be downloaded via \code{\link{getMetaGenomes}}.
}
\examples{
\dontrun{
listMetaGenomes()
} 
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{getMetaGenomes}}, \code{\link{getMetaGenomeSummary}}
}

