% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSubgroups.R
\name{listSubgroups}
\alias{listSubgroups}
\title{List number of available genomes in each group}
\usage{
listSubgroups(db = "refseq")
}
\arguments{
\item{db}{a character string specifying the database for which genome availability shall be checked, 
e.g. \code{db = "refseq"} and \code{db = "genbank"}.}
}
\description{
Users can retrieve the available number of sequenced genomes per group. 
Only available for \code{db = "refseq"} and \code{db = "genbank"}.
}
\examples{
\dontrun {
# example for refseq
listSubgroups(db = "refseq")

# example for genbank
listSubgroups(db = "genbank")
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{listGenomes}}, \code{\link{is.genome.available}}, \code{\link{listKingdoms}}, \code{\link{listGroups}}
}

