% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phytomine.organisms.R
\name{phytomine.organisms}
\alias{phytomine.organisms}
\title{Retrieve All Organism Names Stored on Phytozome v10}
\usage{
phytomine.organisms(organism = "all")
}
\arguments{
\item{organism}{a character string specifying the scientific name of an organism for which data information shall be retrieved.
Default is \code{organism} = \code{"all"}.}
}
\description{
This function extracts all organism names (scientific names) for which genomes,
proteomes, and CDS files are stored on the Phytozome v10 server.
}
\details{
This function queries the Phytozome v10 API and retrieves all organisms specific information
stored on Phytozome. This way annotation information can be accessed allowing a better
transparency of genome data used for omics studies.
}
\examples{
# retrieve data for all available organisms stored in Phytozome
head(phytomine.organisms())

# retrieve data for Arabidopsis thaliana available organisms stored in Phytozome
phytomine.organisms(organism = "Arabidopsis thaliana")
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://phytozome.jgi.doe.gov/phytomine/begin.do}
}

