% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phytozomeOrganisms.R
\name{phytozomeOrganisms}
\alias{phytozomeOrganisms}
\title{Retrieve All Organism Names Stored on Phytozome v10}
\usage{
phytozomeOrganisms()
}
\description{
This function extracts all organism names (scientific names) for which genomes,
proteomes, and CDS files are stored on the Phytozome v10 server.
}
\details{
This function queries the Phytozome v10 API and retrieves all organisms specific information
stored on Phytozome. This way annotation information can be accessed allowing a better
transparency of genome data used for omics studies.
}
\examples{
# retrieve data for all available organisms stored in Phytozome
head(phytozomeOrganisms())
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://phytozome.jgi.doe.gov/phytomine/begin.do}
}

