% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randomSeqs}
\alias{randomSeqs}
\title{Random sequence generator based on a multinomial model}
\usage{
randomSeqs(seq, sample_size)
}
\arguments{
\item{seq}{a character vector storing a sequence as string for which random sequences shall be computed.}

\item{sample_size}{a numeric value specifying the number of random sequences that shall be returned.}
}
\value{
a character vector of length \code{sample_size} storing the random string objects.
}
\description{
This function computes random sequences based on the alphabet and
word length of an input sequence based on a multinomial model.
}
\details{
This function enables you to create a test statistic for sequence comparisons that are based
on a multinomial model assumption. When spexifying the \code{sample_size} argument
a vector of strings of corresponding \code{sample_size} is being returned.

The random strings returned by \code{randomSeqs} have the same word length and are drawn from
the same alphabet as the input sequence.
}
\examples{
# a nucleotide example
seq_example <- "ACCTGGAATTC"

randomSeqs(seq = seq_example, sample_size = 10)

# a protein example
seq_example <- "NPPAAM"

randomSeqs(seq = seq_example, sample_size = 10)
}
\author{
Hajk-Georg Drost
}

