% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randSeqDistr}
\alias{randSeqDistr}
\title{Function to generate a distribution of alignment scores based on random sequences}
\usage{
randSeqDistr(seq, sampled_strings, FUN, ...)
}
\arguments{
\item{seq}{}

\item{sampled_strings}{}

\item{FUN}{}
}
\value{
a numeric vector storing pairwise alignment scores.
}
\description{
This function takes a sequence, a vector of randomly sampled sequences, and a
alignment function as input and computes an pairwise alignment score distribution from the input sequence
and all random sequences stored in \code{sampled_strings}.
}
\examples{
# pairwise alignments using Biostrings
}
\author{
Hajk-Georg Drost
}

