% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.cds}
\alias{read.cds}
\title{Function to read the CDS of a given organism}
\usage{
read.cds(file, format, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the CDS.}

\item{format}{a character string specifying the file format used to store the CDS, e.g. "fasta", "gbk".}

\item{...}{additional arguments that are used by the seqinr::read.fasta() function.}
}
\value{
A data.table storing the gene id in the first column and the corresponding
sequence as string in the second column.
}
\description{
This function reads an organism specific CDS stored in a defined file format.
}
\details{
The \code{read.cds} function takes a string specifying the path to the cds file
of interest as first argument.

It is possible to read in different proteome file standards such as \emph{fasta} or \emph{genebank}.

CDS stored in fasta files can be downloaded from http://www.ensembl.org/info/data/ftp/index.html.
}
\examples{
\dontrun{
# reading a cds file stored in fasta format
Aly.cds <- read.cds("example_cds.fasta", format = "fasta")
}
}
\author{
Hajk-Georg Drost
}

