% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.genome}
\alias{read.genome}
\title{Function to read a genome of a given organism}
\usage{
read.genome(file, format, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the genome.}

\item{format}{a character string specifying the file format used to store the genome, e.g. "fasta", "gbk".}

\item{...}{additional arguments that are used by the seqinr::read.fasta() function.}
}
\value{
A data.table storing the gene id in the first column and the corresponding
sequence as string in the second column.
}
\description{
This function reads an organism specific genome stored in a defined file format.
}
\details{
The \code{read.genome} function takes a string specifying the path to the genome file
of interest as first argument.

It is possible to read in different genome file standards such as \emph{fasta} or \emph{genebank}.
Genomes stored in fasta files can be downloaded from http://ensemblgenomes.org/info/genomes.
}
\examples{
\dontrun{
# reading a genome stored in a fasta file
Aly.genome <- read.genome("example_dna.fasta", format = "fasta")
}
}
\author{
Hajk-Georg Drost
}

