% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rm.R
\name{read_rm}
\alias{read_rm}
\title{Import Repeat Masker output file}
\usage{
read_rm(file)
}
\arguments{
\item{file}{a character string specifying the path to the file 
storing the Repeat Masker output (e.g. retrieved with \code{\link{getRepeatMasker}}).}
}
\description{
This function reads an organism specific 
Repeat Masker output file.
}
\details{
This function takes a string specifying the path to the 
Repeat Masker output file of interest as first argument.
}
\seealso{
\code{\link{getRepeatMasker}}, \code{\link{read_genome}}, 
\code{\link{read_proteome}}, \code{\link{read_gff}}, \code{\link{read_rna}}
}
\author{
Hajk-Georg Drost
}
