% Generated by roxygen2 (4.0.1): do not edit by hand
\name{retrieve_sequence}
\alias{retrieve_sequence}
\title{Helper function for geneSequence()}
\usage{
retrieve_sequence(gene)
}
\arguments{
\item{gene}{a character vector storing the gene id of a organisms of interest to be queried against the
acnucdb database.}
}
\description{
This function takes a gene id as character and returns the biological sequence
of the corresponding gene id.
}
\details{
Sequence information is retrieved from the acnucdb database.
}
\author{
Hajk-Georg Drost
}

