% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cds.R
\name{summary_cds}
\alias{summary_cds}
\title{Retrieve summary statistics for a coding sequence (CDS) file}
\usage{
summary_cds(file, organism)
}
\arguments{
\item{file}{file path to a CDS file in \code{fasta} format.}

\item{organism}{character string specifying the organism at hand.}
}
\description{
A summary statistics of specific CDS features is returned.
}
\details{
The summary statistics include:
\itemize{
\item \code{total_seqs}: 
\item \code{nnn_abs}: The total number of NNN's 
(over all chromosomes/scaffolds/contigs) in all coding sequences combined
\item \code{nnn_perc}: The percentage (relative frequency) of NNN's 
(over all chromosomes/scaffolds/contigs) compared to the total number of 
nucleotides of all coding sequences
}
}
\seealso{
\code{\link{getCollection}}, \code{\link{getCDS}}, \code{\link{read_cds}}, \code{\link{summary_genome}}
}
\author{
Hajk-Georg Drost
}
