% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genome.R
\name{summary_genome}
\alias{summary_genome}
\title{Retrieve summary statistics for a genome assembly file}
\usage{
summary_genome(file, organism)
}
\arguments{
\item{file}{file path to a genome assembly file in \code{fasta} format.}

\item{organism}{character string specifying the organism at hand.}
}
\description{
A summary statistics of specific genome features is returned.
}
\details{
The summary statistics include:
\itemize{
\item \code{genome_size_mbp}: Genome size in mega base pairs
\item \code{n50_mbp}: The N50 contig size of the genome assembly in mega base pairs
\item \code{n_seqs}: The number of chromosomes/scaffolds/contigs of the genome assembly file
\item \code{n_nnn}: The absolute number of NNN's 
(over all chromosomes/scaffolds/contigs) in the genome assembly file
\item \code{rel_nnn}: The percentage (relative frequency) of NNN's 
(over all chromosomes/scaffolds/contigs) compared to the total number of 
nucleotides in the genome assembly file
\item \code{genome_entropy}: The \code{Shannon Entropy} of the genome assembly file
(median entropy over all individual chromosome entropies)}
\item \code{n_gc}: The total number of GC's 
(over all chromosomes/scaffolds/contigs) in the genome assembly file
\item \code{rel_gc}: The (relative frequency) of GC's 
(over all chromosomes/scaffolds/contigs) compared to the total number of 
nucleotides in the genome assembly file
}
\seealso{
\code{\link{getCollection}}, \code{\link{getGenome}}, \code{\link{read_genome}}
}
\author{
Hajk-Georg Drost
}
