% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visGenome}
\alias{visGenome}
\title{Visualization function for genome properties}
\usage{
visGenome(data, type = "length_distr", ...)
}
\arguments{
\item{data}{a data.table object storing the corresponding genome as returned by \code{\link{read.genome}}, \code{\link{read.proteome}}, or
\code{\link{read.cds}}.}

\item{type}{a character string specifying the genome feature that shall be visualized: "length_distr", ... .}

\item{...}{additional parameters that shall be passed to the corresponding plot function.}
}
\value{
a plot object.
}
\description{
This function allows to visualize selected properties of input genomes such as:
\itemize{
\item gene length distribution: \code{type} = \code{"length_distr"}

}
}
\examples{
\dontrun{

# download the proteome of Arabidopsis thaliana from refseq
# and store the corresponding proteome file in '_ncbi_downloads/proteomes'
Ath_proteome <- getProteome(db = "refseq", kingdom = "plant",
                            organism = "Arabidopsis thaliana",
                            clean_folder = FALSE)

# visualize the length distribution of the proteome of A. thaliana
visGenome(data = Ath_proteome, type = "length_distr")

}
}
\author{
Hajk-Georg Drost
}

