% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zebrafishmine.organisms.R
\name{zebrafishmine.organisms}
\alias{zebrafishmine.organisms}
\title{Retrieve All Organism Names Stored on ZebrafishMine}
\usage{
zebrafishmine.organisms(organism = "all")
}
\arguments{
\item{organism}{a character string specifying the scientific name of an organism for which data information shall be retrieved.
Default is \code{organism} = \code{"all"}.}
}
\description{
This function extracts all organism names (scientific names) for which genomes,
proteomes, and CDS files are stored on the ZebrafishMine server.
}
\details{
This function queries the ZebrafishMine API and retrieves all organisms specific information
stored on ZebrafishMine. This way annotation information can be accessed allowing a better
transparency of genome data used for omics studies.
}
\examples{
# retrieve data for all available organisms stored in ZebrafishMine
head(zebrafishmine.organisms())

# retrieve data for Arabidopsis thaliana available organisms stored in ZebrafishMine
zebrafishmine.organisms(organism = "Danio rerio")
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://zebrafishmine.org/begin.do}
}

