% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneSequence.R
\name{geneSequence}
\alias{geneSequence}
\title{Retrieve Biological Sequences of a given Set of Genes}
\usage{
geneSequence(genes, db)
}
\arguments{
\item{genes}{a character vector storing the gene id of a organisms of interest to be queried against the
acnucdb database.}

\item{db}{a character string specifying the database from which the sequence shall be retrieved.}
}
\value{
A list of sequences for each corresponding input gene stored as string.
}
\description{
This function takes an character vector storing gene ids of interest
and returns the biological sequence of the corresponding gene ids.
}
\details{
Sequence information is retrieved from the acnucdb database.
}
\examples{
\dontrun{

# retrieve amino acid sequences from the 'swissprot' database
# for gene ids:"AT1G06090" = "Q9LND9" and "AT1G06100" = "Q9LND8"

 seqs <- geneSequence(c("Q9LND9","Q9LND8"), db = "swissprot")

# choose different databases available

   choosebank()
}
}
\author{
Hajk-Georg Drost
}

