% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getENSEMBLInfo.R
\name{getENSEMBLInfo}
\alias{getENSEMBLInfo}
\title{Retrieve ENSEMBL info file}
\usage{
getENSEMBLInfo()
}
\value{
a tibble table with info
}
\description{
Retrieve species and genome information from
http://rest.ensembl.org/info/species?content-type=application/json/.
}
\author{
Hajk-Georg Drost
}
