% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGENOMEREPORT.R
\name{getGENOMEREPORT}
\alias{getGENOMEREPORT}
\title{Retrieve NCBI GENOME_REPORTS file}
\usage{
getGENOMEREPORT()
}
\description{
Retrieves NCBI GENOME_REPORTS file from 
ftp://ftp.ncbi.nlm.nih.gov/genomes/GENOME_REPORTS/overview.txt.
}
\author{
Hajk-Georg Drost
}
