% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGO}
\alias{getGO}
\title{Function to retrieve GO terms for a given set of genes from a query organism}
\usage{
getGO(organism, genes, filters, ...)
}
\arguments{
\item{organism}{a character string specifying the scientific name of a query organism.}

\item{genes}{a character vector storing the gene ids of a organisms of interest to be queried against BioMart.}

\item{filters}{a character vector specifying the filter (query key) for the BioMart query, e.g. \code{filter} = \code{"ensembl_gene_id"}.}

\item{...}{additional parameters for the \code{\link{biomart}} function.}
}
\description{
This function takes a gene id as character vector from a given query organism and returns the
corresponding GO terms and additional GO information.
}
\details{
This function takes the scientific name of a query organism, a set of genes for which GO terms
and additional information shall be retrieved, and a filter argument that specifies the attribute
for the query genes.
}
\examples{
\dontrun{

GO_tbl <- getGO(organism = "Arabidopsis thaliana",
                genes    = c("AT1G06090", "AT1G06100", "AT1G06110", "AT1G06120", "AT1G06130", "AT1G06200"),
                filters  = "tair_locus")

# look at the result
head(GO_tbl[ , c("tair_locus","go_accession","goslim_goa_accession","goslim_goa_description")])

}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{biomart}}, \code{\link{organismFilters}}, \code{\link{organismBM}}, \code{\link[biomaRt]{getBM}}
}

