% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGO}
\alias{getGO}
\title{Function to retrieve GO terms for a given set of genes}
\usage{
getGO(...)
}
\arguments{
\item{...}{}
}
\description{
This function takes a gene id as character vector and returns the
corresponding GO term.
}
\details{
GO.
}
\examples{
\dontrun{
getGO(genes =c("AT1G06090", "AT1G06100", "AT1G06110", "AT1G06120", "AT1G06130", "AT1G06200"),
mart = "plants_mart_22",dataset = "athaliana_eg_gene",
attributes = c("go_accession","go_definition_1006","go_name_1006"),
filters = "tair_locus",uniqueRows = FALSE)

}
}
\author{
Hajk-Georg Drost
}

