% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{Retrieve available groups for a kingdom of life}
\usage{
getGroups(kingdom, db = "refseq")
}
\arguments{
\item{kingdom}{a character string specifying for which kingdom of life groups shall be retrieved. See \code{\link{getKingdoms}} for details.}

\item{db}{a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}.
Default is \code{db = "refseq"}.}
}
\description{
A short list of available groups for a kingdom of life.
}
\examples{
\dontrun{ 
# get possible kigdom names
getKingdoms(db = "refseq")
# retrieve subgroups for vertebrate_mammalian available from refseq
getGroups(kingdom = "vertebrate_mammalian", db = "refseq")

# get possible kigdom names
getKingdoms(db = "genbank")
# retrieve subgroups for vertebrate_mammalian available from genbank
getGroups(kingdom = "vertebrate_mammalian",db = "genbank")
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}, \code{\link{getKingdoms}}
}

