% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKingdomAssemblySummary.R
\name{getKingdomAssemblySummary}
\alias{getKingdomAssemblySummary}
\title{Retrieve and summarise the assembly_summary.txt files from NCBI for all kingdoms}
\usage{
getKingdomAssemblySummary(db)
}
\arguments{
\item{db}{database name. E.g. \code{refseq} or \code{genbank}.}
}
\description{
Retrieval function of the assembly_summary.txt file from NCBI for all kingdoms.
The assembly_summary.txt files store available species on NCBI.
}
\examples{
\dontrun{
test <- getKingdomAssemblySummary(db = "refseq")
test
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{getSummaryFile}}, \code{\link{getMetaGenomeSummary}}
}

