% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKingdoms.R
\name{getKingdoms}
\alias{getKingdoms}
\title{Retrieve available kingdoms of life stored in databases}
\usage{
getKingdoms(db = "refseq")
}
\arguments{
\item{db}{a character string specifying the database from which the genome shall be retrieved: \code{db = "refseq"}, \code{db = "genbank"}, \code{db = "ensembl"}, \code{db = "ensemblgenomes"}.
Default is \code{db = "refseq"}.}
}
\description{
A short list of kingdoms of life that are stored in the RefSeq
database and that can be downloaded using e.g. \code{\link{meta.retrieval}}, \code{\link{getGenome}}, etc.
}
\examples{
# retrieve kingdoms available at refseq
getKingdoms(db = "refseq")

# retrieve kingdoms available at genbank
getKingdoms(db = "genbank")
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{meta.retrieval}}, \code{\link{getGenome}}, \code{\link{getProteome}}, \code{\link{getCDS}}
}

