% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarts.R
\name{getMarts}
\alias{getMarts}
\title{Retrieve information about available Ensembl Biomart databases}
\usage{
getMarts()
}
\description{
This funcion queries the Ensembl Biomart API and returns a table
storing information about all available Ensembl Biomart databases.
}
\examples{
# get a table of all available databases from Ensembl Biomart
 getMarts()
}
\seealso{
\code{\link{getDatasets}}, \code{\link{getAttributes}}, 
\code{\link{getFilters}}, \code{\link{organismBM}}, 
\code{\link{organismFilters}}, \code{\link{organismAttributes}}
}
\author{
Hajk-Georg Drost
}
