% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarts.R
\name{getMarts}
\alias{getMarts}
\title{Retrieve All Available BioMart Databases}
\usage{
getMarts()
}
\description{
This funcion queries the BioMart API and returns a table
storing all available BioMart databases.
}
\examples{

\dontrun{
# get a table of all available databases from BioMart
 getMarts()
}

}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{getDatasets}}, \code{\link{getAttributes}}, \code{\link{getFilters}}, \code{\link{organismBM}}, \code{\link{organismFilters}}, \code{\link{organismAttributes}}
}

