% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaGenomeSummary.R
\name{getMetaGenomeSummary}
\alias{getMetaGenomeSummary}
\title{Retrieve the assembly_summary.txt file from NCBI genbank metagenomes}
\usage{
getMetaGenomeSummary()
}
\description{
Retrieval function of the assembly_summary.txt file 
from NCBI genbank metagenomes.
This files stores all available metagenome projects on NCBI Genbank.
}
\examples{
test <- getMetaGenomeSummary()
test
}
\seealso{
\code{\link{getKingdomAssemblySummary}}, 
\code{\link{getSummaryFile}}
}
\author{
Hajk-Georg Drost
}
