% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaGenomeSummary.R
\name{getMetaGenomeSummary}
\alias{getMetaGenomeSummary}
\title{Retrieve the assembly_summary.txt file from NCBI genbank metagenomes}
\usage{
getMetaGenomeSummary(
  local_file = file.path(cachedir(), "assembly_summary_metagenomes_genbank.txt")
)
}
\description{
Retrieval function of the assembly_summary.txt file
from NCBI genbank metagenomes.
This files stores all available metagenome projects on NCBI Genbank.
}
\examples{
\dontrun{
meta.summary <- getMetaGenomeSummary()
meta.summary
}
}
\seealso{
\code{\link{getKingdomAssemblySummary}},
\code{\link{getSummaryFile}}
}
\author{
Hajk-Georg Drost
}
