% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDatabases.R
\name{listDatabases}
\alias{listDatabases}
\title{Retrieve a List of Available NCBI Databases for Download}
\usage{
listDatabases(db_name = "nr", update = FALSE)
}
\arguments{
\item{db_name}{a character string specifying the name of the database that shall be searched for.}

\item{update}{a logical value specifying whether or not the local listDatabases.txt file shall be updated by remote access to NCBI.}
}
\description{
This function allows you to retrieve a list of database names and versions
that can be downloaded from correspondning servers.

Database retrieval is crucial for most biological studies and analyses.
There is a vast diversity of databases that can be accessed remotely or that can be downloaded
to your local machine. This function provides an interface to databases that can be downloaded 
from NCBI servers and lists all available databases and their database version to be able to
select an appropriate database for download with \code{\link{download_database}}.
}
\examples{

# retrieve all versions of the NCBI 'nr' database that can be downloaded
# listDatabases(db_name = "nr")

# analogous:
# listDatabases(db_name = "cdd")
# listDatabases(db_name = "nt")
# listDatabases(db_name = "gss")
# listDatabases(db_name = "refseq_protein")

}
\author{
Hajk-Georg Drost
}
\references{
\url{ftp://ftp.ncbi.nlm.nih.gov/blast/db/FASTA}
}
\seealso{
\code{\link{download_database}}
}

