% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.retrieval.R
\name{meta.retrieval}
\alias{meta.retrieval}
\title{Perform Meta-Genome Retieval}
\usage{
meta.retrieval(kingdom, type = "genome", out.folder = NULL)
}
\arguments{
\item{kingdom}{a character string specifying the kingdom of the organisms of interest,
e.g. "archaea","bacteria", "fungi", "invertebrate", "plant", "protozoa", "vertebrate_mammalian", or "vertebrate_other".}

\item{type}{type of sequences that shall be retrieved. Either \code{genome}, \code{proteome}, or \code{CDS}.}

\item{out.folder}{path to the folder in which downloaded genomes shall be stored. By default the
kingdom name is used to name the output folder.}
}
\description{
Download genomes, proteomes, or CDS of all species within a kingdom of life.
}
\details{
This function aims to perform bulk retrieval of the genomes of species
that belong to the same kingdom of life.
}
\examples{
\dontrun{
# download all vertebrate genomes
meta.retrieval(kingdom = "vertebrate_mammalian", type = "genome")
}
}
\author{
Hajk-Georg Drost
}

