% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randSeqDistr.R
\name{randSeqDistr}
\alias{randSeqDistr}
\title{Generate a Distribution of Alignment Scores Based on Random Sequences}
\usage{
randSeqDistr(seq, subject, sampled_strings, FUN, ..., comp_cores = 1)
}
\arguments{
\item{seq}{a character vector storing a sequence as string for which random sequences shall be computed.}

\item{subject}{a character vector storing a subject sequence as string to which \code{seq} shall be pairwise aligned.}

\item{sampled_strings}{an vector object returned by the \code{\link{randomSeqs}} function.}

\item{FUN}{a pairwise alignment function such as \code{\link[Biostrings]{pairwiseAlignment}} or any other function
that takes sequence arguments as first and second input.}

\item{...}{additional arguments that shall be passed to \code{FUN}.}

\item{comp_cores}{a numeric value specifying the number of cores you want to use for multicore processing.}
}
\value{
a numeric vector storing pairwise alignment scores.
}
\description{
This function takes a sequence, a vector of randomly sampled sequences, and a
alignment function as input and computes an pairwise alignment score distribution from the input sequence
and all random sequences stored in \code{sampled_strings}.
}
\examples{
# pairwise alignments using Biostrings::pairwiseAlignment() function
# you can also specify the comp_cores argument in case you
# are working with a multicore machine

seq_example <- "MEDQVGFGF"
subject_example <- "AYAIDPTPAF"

randScores <- randSeqDistr(seq_example,subject_example, randomSeqs(seq_example,10), Biostrings::pairwiseAlignment, scoreOnly=TRUE, comp_cores = 1)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{evalAlignment}}, \code{\link{randomSeqs}}
}

