% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refseqOrganisms.R
\name{refseqOrganisms}
\alias{refseqOrganisms}
\title{Retrieve All Organism Names Stored on refseq}
\usage{
refseqOrganisms()
}
\description{
This function extracts all organism names (scientific names) for which genomes,
proteomes, and CDS files are stored on the NCBI refseq server.
}
\author{
Hajk-Georg Drost
}
\references{
\url{ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/}
}

