% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genome.R
\name{summary_genome}
\alias{summary_genome}
\title{Retrieve summary statistics for a genome assembly file}
\usage{
summary_genome(file, organism)
}
\arguments{
\item{file}{file path to a genome assembly file in \code{fasta} format.}

\item{organism}{character string specifying the organism at hand.}
}
\description{
A summary statistics of specific genome features is returned.
}
\seealso{
\code{\link{getCollection}}, \code{\link{getGenome}}, \code{\link{read_genome}}
}
\author{
Hajk-Georg Drost
}
