% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visGenome}
\alias{visGenome}
\title{Visualization function for genome properties}
\usage{
visGenome(data, type = "length_distr", ...)
}
\arguments{
\item{data}{a data.table object storing the corresponding genome as returned by \code{\link{read.genome}}, \code{\link{read.proteome}}, or
\code{\link{read.cds}}.}

\item{type}{a character string specifying the genome feature that shall be visualized: "length_distr", ... .}

\item{...}{additional parameters that shall be passed to the corresponding plot function.}
}
\value{
a plot object.
}
\description{
This function allows to visualize selected properties of input genomes such as:
\itemize{
\item gene length distribution

}
}
\author{
Hajk-Georg Drost
}

